split assertions test into multiple smaller test classes
