+++ junit-jupiter-engine/src/test/java/org/junit/jupiter/api/AssertionsAssertAllTests.java
@@ -1,0 +1,141 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.io.IOException;
+import java.util.List;
+import java.util.stream.Stream;
+
+import org.junit.jupiter.api.function.Executable;
+import org.junit.platform.commons.util.PreconditionViolationException;
+import org.opentest4j.AssertionFailedError;
+import org.opentest4j.MultipleFailuresError;
+
+/**
+ * Unit tests for JUnit Jupiter {@link Assertions}.
+ *
+ * @since 5.0
+ */
+public class AssertionsAssertAllTests implements AssertionsHelper {
+
+	@Test
+	void assertAllWithNullInExecutableArray() {
+		try {
+			// @formatter:off
+			assertAll(
+				() -> {},
+				(Executable) null
+			);
+			// @formatter:on
+		}
+		catch (PreconditionViolationException ex) {
+			assertMessageEquals(ex, "individual executables must not be null");
+		}
+	}
+
+	@Test
+	void assertAllWithNullExecutableArray() {
+		try {
+			assertAll((Executable[]) null);
+		}
+		catch (PreconditionViolationException ex) {
+			assertMessageEquals(ex, "executables array must not be null or empty");
+		}
+	}
+
+	@Test
+	void assertAllWithNullExecutableStream() {
+		try {
+			assertAll((Stream<Executable>) null);
+		}
+		catch (PreconditionViolationException ex) {
+			assertMessageEquals(ex, "executables must not be null");
+		}
+	}
+
+	@Test
+	void assertAllWithExecutablesThatDoNotThrowExceptions() {
+		// @formatter:off
+		assertAll(
+			() -> assertTrue(true),
+			() -> assertFalse(false),
+			() -> assertTrue(true)
+		);
+		// @formatter:on
+	}
+
+	@Test
+	void assertAllWithExecutablesThatThrowAssertionErrors() {
+		// @formatter:off
+		MultipleFailuresError multipleFailuresError = assertThrows(MultipleFailuresError.class, () ->
+			assertAll(
+				() -> assertFalse(true),
+				() -> assertFalse(true)
+			)
+		);
+		// @formatter:on
+
+		assertTrue(multipleFailuresError != null);
+		List<AssertionError> failures = multipleFailuresError.getFailures();
+		assertTrue(failures.size() == 2);
+		assertTrue(failures.get(0).getClass().equals(AssertionFailedError.class));
+	}
+
+	@Test
+	void assertAllWithStreamOfExecutablesThatThrowAssertionErrors() {
+		// @formatter:off
+		MultipleFailuresError multipleFailuresError = assertThrows(MultipleFailuresError.class, () ->
+			assertAll(Stream.of(() -> assertFalse(true), () -> assertFalse(true)))
+		);
+		// @formatter:on
+
+		assertTrue(multipleFailuresError != null);
+		List<AssertionError> failures = multipleFailuresError.getFailures();
+		assertTrue(failures.size() == 2);
+		assertTrue(failures.get(0).getClass().equals(AssertionFailedError.class));
+	}
+
+	@Test
+	void assertAllWithExecutableThatThrowsThrowable() {
+		assertThrows(EnigmaThrowable.class, () -> assertAll(() -> {
+			throw new EnigmaThrowable();
+		}));
+	}
+
+	@Test
+	void assertAllWithExecutableThatThrowsCheckedException() {
+		assertThrows(IOException.class, () -> assertAll(() -> {
+			throw new IOException();
+		}));
+	}
+
+	@Test
+	void assertAllWithExecutableThatThrowsRuntimeException() {
+		assertThrows(IllegalStateException.class, () -> assertAll(() -> {
+			throw new IllegalStateException();
+		}));
+	}
+
+	@Test
+	void assertAllWithExecutableThatThrowsError() {
+		assertThrows(StackOverflowError.class, () -> assertAll(this::recurseIndefinitely));
+	}
+
+	@SuppressWarnings("serial")
+	private static class EnigmaThrowable extends Throwable {
+	}
+
+}
