+++ junit-jupiter-engine/src/test/java/org/junit/jupiter/api/AssertionsAssertFalseTests.java
@@ -1,0 +1,78 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+
+import org.opentest4j.AssertionFailedError;
+
+/**
+ * Unit tests for JUnit Jupiter {@link Assertions}.
+ *
+ * @since 5.0
+ */
+public class AssertionsAssertFalseTests implements AssertionsHelper {
+
+	@Test
+	void assertFalseWithBooleanFalse() {
+		assertFalse(false);
+	}
+
+	@Test
+	void assertFalseWithBooleanSupplierFalse() {
+		assertFalse(() -> false);
+	}
+
+	@Test
+	void assertFalseWithBooleanTrueAndString() {
+		try {
+			assertFalse(true, "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "test");
+		}
+	}
+
+	@Test
+	void assertFalseWithBooleanTrueAndMessageSupplier() {
+		try {
+			assertFalse(true, () -> "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "test");
+		}
+	}
+
+	@Test
+	void assertFalseWithBooleanSupplierTrueAndString() {
+		try {
+			assertFalse(() -> true, "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "test");
+		}
+	}
+
+	@Test
+	void assertFalseWithBooleanSupplierTrueAndMessageSupplier() {
+		try {
+			assertFalse(() -> true, () -> "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "test");
+		}
+	}
+
+}
