+++ junit-jupiter-engine/src/test/java/org/junit/jupiter/api/AssertionsAssertNotNullTests.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import org.opentest4j.AssertionFailedError;
+
+/**
+ * Unit tests for JUnit Jupiter {@link Assertions}.
+ *
+ * @since 5.0
+ */
+public class AssertionsAssertNotNullTests implements AssertionsHelper {
+
+	@Test
+	void assertNotNullWithNonNullObject() {
+		assertNotNull("foo");
+	}
+
+	@Test
+	void assertNotNullWithNull() {
+		try {
+			assertNotNull(null);
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "expected: not <null>");
+		}
+	}
+
+	@Test
+	void assertNotNullWithNullAndMessageSupplier() {
+		try {
+			assertNotNull(null, () -> "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageStartsWith(ex, "test");
+			assertMessageEndsWith(ex, "expected: not <null>");
+		}
+	}
+
+}
