+++ junit-jupiter-engine/src/test/java/org/junit/jupiter/api/AssertionsAssertNullTests.java
@@ -1,0 +1,67 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Assertions.assertNull;
+
+import org.opentest4j.AssertionFailedError;
+
+/**
+ * Unit tests for JUnit Jupiter {@link Assertions}.
+ *
+ * @since 5.0
+ */
+public class AssertionsAssertNullTests implements AssertionsHelper {
+
+	@Test
+	void assertNullWithNull() {
+		assertNull(null);
+	}
+
+	@Test
+	void assertNullWithNonNullObject() {
+		try {
+			assertNull("foo");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEndsWith(ex, "expected: <null> but was: <foo>");
+			assertExpectedAndActualValues(ex, null, "foo");
+		}
+	}
+
+	@Test
+	void assertNullWithNonNullObjectAndMessage() {
+		try {
+			assertNull("foo", "a message");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageStartsWith(ex, "a message");
+			assertMessageEndsWith(ex, "expected: <null> but was: <foo>");
+			assertExpectedAndActualValues(ex, null, "foo");
+		}
+	}
+
+	@Test
+	void assertNullWithNonNullObjectAndMessageSupplier() {
+		try {
+			assertNull("foo", () -> "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageStartsWith(ex, "test");
+			assertMessageEndsWith(ex, "expected: <null> but was: <foo>");
+			assertExpectedAndActualValues(ex, null, "foo");
+		}
+	}
+
+}
