+++ junit-jupiter-engine/src/test/java/org/junit/jupiter/api/AssertionsFailTests.java
@@ -1,0 +1,70 @@
+/*
+ * Copyright 2015-2016 the original author or authors.
+ *
+ * All rights reserved. This program and the accompanying materials are
+ * made available under the terms of the Eclipse Public License v1.0 which
+ * accompanies this distribution and is available at
+ *
+ * http://www.eclipse.org/legal/epl-v10.html
+ */
+
+package org.junit.jupiter.api;
+
+import static org.junit.jupiter.api.Assertions.fail;
+
+import java.util.function.Supplier;
+
+import org.opentest4j.AssertionFailedError;
+
+/**
+ * Unit tests for JUnit Jupiter {@link Assertions}.
+ *
+ * @since 5.0
+ */
+public class AssertionsFailTests implements AssertionsHelper {
+
+	@Test
+	void failWithString() {
+		try {
+			fail("test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "test");
+		}
+	}
+
+	@Test
+	void failWithMessageSupplier() {
+		try {
+			fail(() -> "test");
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "test");
+		}
+	}
+
+	@Test
+	void failWithNullString() {
+		try {
+			fail((String) null);
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "");
+		}
+	}
+
+	@Test
+	void failWithNullMessageSupplier() {
+		try {
+			fail((Supplier<String>) null);
+			expectAssertionFailedError();
+		}
+		catch (AssertionFailedError ex) {
+			assertMessageEquals(ex, "");
+		}
+	}
+
+}
